debuged print rows sequence
