+++ internal/helpers.go
@@ -428,52 +428,3 @@ func ZipXmlFileDecodeData(zf *zip.File, enc FileEncodingNumber) (*bytes.Reader, 
 	}
 	return breader, err
 }
-
-func PrintLinks(name string, link *LinkedRow) {
-	// TODO: create test instead
-	lnk := link
-	fmt.Println()
-	slog.Debug(name, "output", "printing input")
-	fmt.Println(name, lnk)
-	slog.Debug(name, "output", "printing start")
-	fmt.Println(name, lnk.Start)
-	start := *lnk.Start
-	fmt.Println(name, "start", start)
-	slog.Debug(name, "output", "printing end")
-	fmt.Println(name, "end", *lnk.End)
-	fmt.Println()
-	slog.Debug(name, "output", "printing by previous")
-	count := 10
-	for i := 0; i < count; i++ {
-		fmt.Println(name, i, "seq_prev", lnk)
-		fmt.Println(name, i, "start_prev", *lnk.Start)
-		fmt.Println(name, i, "end_prev", *lnk.End)
-		lnktest := lnk.PrevL
-		if lnktest == nil {
-			slog.Debug(name, "output", "sequence end")
-			break
-		}
-		lnk = lnk.PrevL
-	}
-	fmt.Println()
-	slog.Debug(name, "output", "printing by next")
-	for i := 0; i < count; i++ {
-		fmt.Println(name, i, "seq_next", lnk)
-		fmt.Println(name, i, "start_next", *lnk.Start)
-		fmt.Println(name, i, "end_next", *lnk.End)
-		lnk = lnk.NextL
-		if lnk == nil {
-			slog.Debug(name, "output", "sequence end")
-			break
-		}
-	}
-}
-
-func PrintLinesSlice(lrows []*LinkedRow) {
-	for _, r := range lrows {
-		fmt.Println("current", r.Payload.CSVrow)
-		if r.PrevL != nil {
-			fmt.Println("previous", r.PrevL.Payload.CSVrow)
-		}
-	}
-}
